:fire: Move FormBuilder to notifier directory
